Refactor readers into separate modules
