+++ src/app.js
@@ -1,8 +1,9 @@
 // Three.js via CDN modules (ESM) and helpers from the examples directory.
 import * as THREE from 'three';
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
-import { STLLoader } from 'three/examples/jsm/loaders/STLLoader.js';
-import { DXFLoader } from './loaders/DXFLoader.js';
+import { loadStl } from './readers/stlReader.js';
+import { loadDxf } from './readers/dxfReader.js';
+import { loadStep } from './readers/stepReader.js';
 
 const viewerEl = document.getElementById('viewer');
 const resultsEl = document.getElementById('results');
@@ -16,14 +17,6 @@ let viewerManager = null;
 const models = [];
 
 let occtModulePromise = null;
-
-const unitToMillimeter = {
-  mm: 1,
-  cm: 10,
-  m: 1000,
-  in: 25.4,
-  ft: 304.8,
-};
 
 const PLANAR_ANGLE_THRESHOLD_DEG = 3;
 const CIRCULARITY_THRESHOLD = 0.80;
@@ -82,15 +75,29 @@ async function handleFiles(fileList) {
       if (loading) {
         loading.classList.add('show');
       }
+      const readerOptions = {
+        stlUnitEl,
+        precisionEl,
+        addCard,
+        updateCardBody,
+        computeBoundsFromPositions,
+        computeBoundsFromGroup,
+        dimsFromBounds,
+        analyzeSheetMetal,
+        formatDims,
+        formatLaserCutAnalysis,
+        models,
+        ensureOcctModule,
+      };
       if (lower.endsWith('.stl')) {
         viewport = viewerManager.createViewport(name);
-        await loadStl(file, card, viewport);
+        await loadStl(file, card, viewport, readerOptions);
       } else if (lower.endsWith('.step') || lower.endsWith('.stp')) {
         viewport = viewerManager.createViewport(name);
-        await loadStep(file, card, viewport);
+        await loadStep(file, card, viewport, readerOptions);
       } else if (lower.endsWith('.dxf')) {
         viewport = viewerManager.createViewport(name);
-        await loadDxf(file, card, viewport);
+        await loadDxf(file, card, viewport, readerOptions);
       } else {
         updateCardBody(card, '<div class="warn">Unsupported file type.</div>');
       }
@@ -107,201 +114,6 @@ async function handleFiles(fileList) {
       }
     }
   }
-}
-
-async function loadStl(file, card, viewport) {
-  if (!viewport) {
-    throw new Error('A viewer window could not be created for this STL file.');
-  }
-
-  const arrayBuffer = await file.arrayBuffer();
-  const loader = new STLLoader();
-  const geometry = loader.parse(arrayBuffer);
-
-  const material = new THREE.MeshStandardMaterial({
-    color: 0x9fb5ff,
-    metalness: 0.1,
-    roughness: 0.75,
-  });
-
-  geometry.computeVertexNormals();
-
-  const mesh = new THREE.Mesh(geometry, material);
-  const group = new THREE.Group();
-  group.add(mesh);
-
-  const positionAttr = geometry.getAttribute('position');
-  const positions = positionAttr && positionAttr.array;
-  if (!positions) {
-    throw new Error('STL geometry missing vertex positions.');
-  }
-
-  const unit = (stlUnitEl && stlUnitEl.value) || 'mm';
-  const toMillimeter = Object.prototype.hasOwnProperty.call(unitToMillimeter, unit)
-    ? unitToMillimeter[unit]
-    : 1;
-
-  mesh.scale.setScalar(toMillimeter);
-
-  const bounds = computeBoundsFromPositions(positions);
-  bounds.min.multiplyScalar(toMillimeter);
-  bounds.max.multiplyScalar(toMillimeter);
-
-  const dims = dimsFromBounds(bounds);
-  const analysis = analyzeSheetMetal(group);
-
-  const name = `${file.name}`;
-  const precisionValue = precisionEl && precisionEl.value !== undefined ? precisionEl.value : '3';
-  const decimals = parseInt(precisionValue, 10) || 3;
-  const bodyHtml = [
-    `<div class="ok">Loaded STL (${unit} → mm).</div>`,
-    formatDims(dims, decimals),
-    formatLaserCutAnalysis(analysis, decimals),
-  ].join('');
-  const targetCard = card || addCard(name, bodyHtml);
-  updateCardBody(targetCard, bodyHtml);
-  targetCard.classList.remove('pending');
-
-  viewport.setTitle(name);
-  viewport.setModel(group, bounds.clone());
-
-  const model = { name, group, bounds: bounds.clone(), unit: unit || 'mm', kind: 'stl', viewport, card: targetCard };
-  targetCard.addEventListener('click', () => viewport.focus());
-  models.push(model);
-  return targetCard;
-}
-
-async function loadDxf(file, card, viewport) {
-  if (!file) {
-    throw new Error('A file must be provided to loadDxf.');
-  }
-  if (!viewport) {
-    throw new Error('A viewer window could not be created for this DXF file.');
-  }
-
-  const text = await file.text();
-  const loader = new DXFLoader();
-  const group = loader.parse(text);
-
-  if (!group) {
-    throw new Error('DXF loader produced no geometry.');
-  }
-
-  const bounds = computeBoundsFromGroup(group);
-  if (!bounds) {
-    throw new Error('DXF file contained no drawable entities.');
-  }
-
-  const dimsMm = dimsFromBounds(bounds);
-  const analysis = analyzeSheetMetal(group);
-
-  const name = `${file.name}`;
-  const precisionValue = precisionEl && precisionEl.value !== undefined ? precisionEl.value : '3';
-  const decimals = parseInt(precisionValue, 10) || 3;
-  const bodyHtml = [
-    '<div class="ok">Loaded DXF (flat pattern, units assumed mm).</div>',
-    formatDims(dimsMm, decimals),
-    formatLaserCutAnalysis(analysis, decimals),
-  ].join('');
-  const targetCard = card || addCard(name, bodyHtml);
-  updateCardBody(targetCard, bodyHtml);
-  targetCard.classList.remove('pending');
-
-  viewport.setTitle(name);
-  viewport.setModel(group, bounds.clone());
-
-  const model = { name, group, bounds: bounds.clone(), unit: 'mm', kind: 'dxf', viewport, card: targetCard };
-  targetCard.addEventListener('click', () => viewport.focus());
-  models.push(model);
-  return targetCard;
-}
-
-async function loadStep(file, card, viewport) {
-  if (!file) {
-    throw new Error('A file must be provided to loadStep.');
-  }
-  if (!viewport) {
-    throw new Error('A viewer window could not be created for this STEP file.');
-  }
-
-  const occt = await ensureOcctModule();
-  const uint8 = new Uint8Array(await file.arrayBuffer());
-  const params = {
-    linearUnit: 'millimeter',
-    linearDeflectionType: 'bounding_box_ratio',
-    linearDeflection: 0.00005,
-    angularDeflection: 0.05,
-  };
-  const result = occt.ReadStepFile(uint8, params);
-  if (!result || !result.success) {
-    throw new Error('STEP import failed.');
-  }
-
-  const group = new THREE.Group();
-  let bounds = null;
-  for (const meshResult of result.meshes) {
-    const attr = meshResult && meshResult.attributes;
-    const positionAttr = attr && attr.position;
-    const posSrc = (positionAttr && positionAttr.array)
-      || meshResult.position
-      || meshResult.positions
-      || meshResult.vertices;
-    if (!posSrc) continue;
-
-    const posArray = posSrc.BYTES_PER_ELEMENT ? posSrc : new Float32Array(posSrc);
-
-    const geometry = new THREE.BufferGeometry();
-    geometry.setAttribute('position', new THREE.Float32BufferAttribute(posArray, 3));
-
-    const indexAttr = meshResult.index && meshResult.index.array ? meshResult.index.array : null;
-    const idxSrc = indexAttr || meshResult.indices || meshResult.index;
-    if (idxSrc) {
-      const indexArray = idxSrc.length > 65535 ? new Uint32Array(idxSrc) : new Uint16Array(idxSrc);
-      geometry.setIndex(new THREE.BufferAttribute(indexArray, 1));
-    }
-    geometry.computeVertexNormals();
-
-    const colorObj = (meshResult.color && meshResult.color.length === 3)
-      ? new THREE.Color(meshResult.color[0], meshResult.color[1], meshResult.color[2])
-      : new THREE.Color(0x3c8bff);
-    const material = new THREE.MeshStandardMaterial({
-      color: colorObj,
-      metalness: 0.05,
-      roughness: 0.6,
-    });
-    const mesh = new THREE.Mesh(geometry, material);
-    group.add(mesh);
-
-    const meshBounds = computeBoundsFromPositions(posArray);
-    bounds = bounds ? bounds.union(meshBounds) : meshBounds.clone();
-  }
-
-  if (!bounds) {
-    throw new Error('No geometry produced from STEP.');
-  }
-
-  const dimsMm = dimsFromBounds(bounds);
-  const analysis = analyzeSheetMetal(group);
-
-  const name = `${file.name}`;
-  const precisionValue = precisionEl && precisionEl.value !== undefined ? precisionEl.value : '3';
-  const decimals = parseInt(precisionValue, 10) || 3;
-  const bodyHtml = [
-    '<div class="ok">Loaded STEP (converted → mm).</div>',
-    formatDims(dimsMm, decimals),
-    formatLaserCutAnalysis(analysis, decimals),
-  ].join('');
-  const targetCard = card || addCard(name, bodyHtml);
-  updateCardBody(targetCard, bodyHtml);
-  targetCard.classList.remove('pending');
-
-  viewport.setTitle(name);
-  viewport.setModel(group, bounds.clone());
-
-  const model = { name, group, bounds: bounds.clone(), unit: 'mm', kind: 'step', viewport, card: targetCard };
-  targetCard.addEventListener('click', () => viewport.focus());
-  models.push(model);
-  return targetCard;
 }
 
 async function ensureOcctModule() {
